ai/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
google-generativeai>=0.7.2
Pillow>=10.3
requests>=2.32
